+++ CHANGELOG.md
@@ -29,6 +29,15 @@ You can check your current version with the following command:
     ```
 
 For more information, see [UP42 Python package description](https://pypi.org/project/up42-py/).
+
+## 1.0.4a3
+
+**May 27, 2024**
+
+- Added `Session` and `WorkspaceId` descriptors to provide access to session and workspace_id from other classes.
+- Renaming `main` module to `base`.
+
+
 ## 1.0.4a2
 
 **May 24, 2024**